+++ src/test/java/service/CarRaceServiceImplTest.java
@@ -8,8 +8,6 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import repository.CarRaceResultRepositoryImpl;
-import util.RandomNumberGenerator;
-import util.RandomNumberGeneratorImpl;
 
 class CarRaceServiceImplTest {
