+++ src/test/java/util/RandomNumberGeneratorTest.java
@@ -2,6 +2,8 @@ package util;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import service.RandomNumberGenerator;
+import service.RandomNumberGeneratorImpl;
 
 class RandomNumberGeneratorTest {
     private final RandomNumberGenerator randomNumberGenerator = new RandomNumberGeneratorImpl();
